Fix partial read issue
It was an issue that when POST request body was fragmented, only first
fragment was read by server and remaining data was skipped. That resulted
in sporadical failure:

E0511 10:29:47.986438 1 utils.go:87] Failed to process the request's data.
Error while unmarshaling data. Details: invalid character '\x00' looking for beginning of value

Now it is fixed and transaction is closed after all data was read.